\name{listDatabases}
\docType{methods}

\alias{generic,listDatabases}
\alias{listDatabases,MeSHDb-method}
\alias{listDatabases}

\title{A function to return the scientific name of package}
\description{This function returns the scientific name of species used in the package.}

\usage{listDatabases(x)}

\arguments{
\item{x}{\code{MeSHDb} object such as \code{MeSH.Mmu.eg.db}}
}
\author{Koki Tsuyuzaki}
\examples{
# library("MeSH.Mmu.eg.db")
# listDatabases(MeSH.Mmu.eg.db)
}