\name{packageName}
\docType{methods}

\alias{generic,packageName}
\alias{packageName,MeSHDb-method}
\alias{packageName}

\title{A function to return the name of package}
\description{This function returns the name of package}

\usage{packageName(x)}

\arguments{
\item{x}{\code{MeSHDb} object such as \code{MeSH.Mmu.eg.db}}
}
\author{Koki Tsuyuzaki}
\examples{
# library("MeSH.Mmu.eg.db")
# packageName(MeSH.Mmu.eg.db)
}